+++ src/test/java/net/kiar/collectorr/config/ConfigLoaderTest.java
@@ -2,11 +2,14 @@ package net.kiar.collectorr.config;
 
 import java.io.FileNotFoundException;
 import java.util.List;
+import java.util.Map;
+import net.kiar.collectorr.config.model.FieldValueMap;
+import net.kiar.collectorr.config.model.RootFieldMap;
 import net.kiar.collectorr.config.model.TopicConfig;
-import net.kiar.collectorr.config.model.TopicConfigMappings;
 import net.kiar.collectorr.config.model.TopicConfigMetric;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import org.junit.jupiter.api.Test;
 
 /**
@@ -77,24 +80,53 @@ topics:
 topics:
 - topic: topic/to/test
   mappings:
-    - label:
-      source: sensorId
-      target: sensorName
-      map:
-        284B5096F0013CBA: Sensor1
+    sensorId:
+      281C5A96F0013CAC:
+        sensorName: Abgasrohr
+        namePrefix: heating
 """
         );
         
         assertNotNull(conf);
         TopicConfig item1 = conf.getTopicsToObserve().get(0);
         assertNotNull(item1);
-        List<TopicConfigMappings> mappings = item1.getMappings();
+        RootFieldMap mappings = item1.getValueMappings();
         assertNotNull(mappings);
         assertEquals(1, mappings.size());
-        TopicConfigMappings mapping = mappings.get(0);
-        assertEquals("sensorId", mapping.getSource());
-        assertEquals("sensorName", mapping.getTarget());
-        assertEquals(1, mapping.getMap().size());
+        FieldValueMap fieldMapping = mappings.get("sensorId");
+        assertEquals(1, fieldMapping.size());
+        
+        Map<String, String> pairs = fieldMapping.values().stream().findAny().get();
+        assertEquals(2, pairs.size());
+        assertEquals("Abgasrohr", pairs.get("sensorName"));
+        assertEquals("heating", pairs.get("namePrefix"));
+        assertNull(pairs.get("sensorId"));
+        
+    }
+    
+    
+    @Test
+    public void testValueMapping() {
+        MappingsConfigLoader conf = MappingsConfigLoader.readContent("""
+topics:
+- topic: topic/to/test
+  valueMappings:
+    sensorId:
+      281C5A96F0013CAC:
+        sensorName: Abgasrohr
+        namePrefix: heating
+
+"""
+        );
+        
+        assertNotNull(conf);
+        TopicConfig topicConfig1 = conf.getTopicsToObserve().get(0);
+        assertNotNull(topicConfig1);
+        RootFieldMap mappings = topicConfig1.getValueMappings();
+        assertNotNull(mappings);
+        assertEquals(1, mappings.size());
+        
+        
         
     }